requirements.txt
Flask==0.11
Flask-Script==2.0.5
flask-mongoengine==0.8.2
jsonschema==2.6.0
elasticsearch==6.3.1
blinker==1.4
python-dotenv==0.10.1
faker==1.0.4
kafka-python==1.3.5
redis==3.0.1
Flask-Mail==0.9.1
requests==2.21.0
rq==1.0
flask_limiter==1.0.1
boto3==1.9.183